+++ ChangeLog.md
@@ -9,6 +9,7 @@ Table:
 - Correctly responds with status 202 on merge with non-existent entity.
 - Properly differentiate between upsert and update in batch merge and replace.
 - Added additional tests via raw REST tests.
+- Correctly deletes a table that is a substring of another table.
 
 ## 2022.06 Version 3.18.0